requirements.txt
flask
ultralytics
mediapipe
# Use opencv-python-headless for servers, it's smaller
opencv-python-headless
numpy
Pillow